+++ src/ElementFinderFactory.php
@@ -1,37 +1,70 @@
 <?php
 
+  declare(strict_types=1);
+
   namespace Xparse\Parser;
 
   use Psr\Http\Message\ResponseInterface;
   use Xparse\ElementFinder\ElementFinder;
   use Xparse\ElementFinder\Helper\StringHelper;
-  use Xparse\Parser\Helper\HtmlEncodingConverter;
-  use Xparse\Parser\Helper\LinkConverter;
+  use Xparse\ExpressionTranslator\ExpressionTranslatorInterface;
+  use Xparse\ExpressionTranslator\XpathExpression;
+  use Xparse\Parser\Helper\EncodingConverterInterface;
+  use Xparse\Parser\Helper\LinkConverterInterface;
+  use Xparse\Parser\Helper\RelativeToAbsoluteLinkConverter;
+  use Xparse\Parser\Helper\ToUtfConverter;
 
-  /**
-   * Create ElementFinder.
-   * Convert charset to UTF-8
-   * Convert relative links to absolute
-   * @package Xparse\Parser\Helper
-   */
   class ElementFinderFactory implements ElementFinderFactoryInterface {
 
     /**
-     * @inheritdoc
-     * @throws \InvalidArgumentException
+     * @var LinkConverterInterface
      */
-    public function create(ResponseInterface $response, $affectedUrl = null) {
-      $html = $response->getBody();
-      $html = StringHelper::safeEncodeStr((string) $html);
-      $contentType = $response->getHeaderLine('content-type');
-      $html = HtmlEncodingConverter::convertToUtf($html, $contentType);
-      $page = new ElementFinder((string) $html);
+    private $linkConverter;
+
+    /**
+     * @var EncodingConverterInterface
+     */
+    private $encodingConverter;
+
+    /**
+     * @var ExpressionTranslatorInterface
+     */
+    private $expressionTranslator;
+
+    /**
+     * @param ExpressionTranslatorInterface|null $expressionTranslator
+     * @param LinkConverterInterface|null $linkConverter
+     * @param EncodingConverterInterface|null $encodingConverter
+     */
+    public function __construct(
+      ExpressionTranslatorInterface $expressionTranslator = null,
+      LinkConverterInterface $linkConverter = null,
+      EncodingConverterInterface $encodingConverter = null
+    ) {
+      if ($expressionTranslator === null) {
+        $this->expressionTranslator = new XpathExpression();
+      }
+      if ($linkConverter === null) {
+        $this->linkConverter = new RelativeToAbsoluteLinkConverter();
+      }
+      if ($encodingConverter === null) {
+        $this->encodingConverter = new ToUtfConverter();
+      }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function create(ResponseInterface $response, string $affectedUrl = '') : ElementFinder {
+      $html = StringHelper::safeEncodeStr($response->getBody()->getContents());
+      $html = $this->encodingConverter->convert($html, $response->getHeaderLine('content-type'));
+      $elementFinder = new ElementFinder($html, null, $this->expressionTranslator);
 
       if ($affectedUrl !== null) {
-        LinkConverter::convertUrlsToAbsolute($page, $affectedUrl);
+        $this->linkConverter->convert($elementFinder, $affectedUrl);
       }
 
-      return $page;
+      return $elementFinder;
     }
 
   }